last commit for this project

+++ src/components/Products/ProductBox.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import React from "react";
+import React, { useEffect } from "react";
 import { useContext, MainContext } from "../../contexts/HomeContext.js";
 
 const ProductBox = ({ price, title, image, id, category }) => {
@@ -15,19 +15,27 @@ const ProductBox = ({ price, title, image, id, category }) => {
       category,
     };
 
-    const updatedBasketItems = [...basketItems, newItem];
-    setBasketItems(updatedBasketItems);
-    localStorage.setItem("basketItems", JSON.stringify(updatedBasketItems));
+    const existingItem = basketItems.find((item) => item.id === id);
+
+    if (existingItem) {
+      console.log(`Item with id ${id} already exists in the basket.`);
+    } else {
+      const updatedBasketItems = [...basketItems, newItem];
+      setBasketItems(updatedBasketItems);
+      localStorage.setItem("basketItems", JSON.stringify(updatedBasketItems));
+    }
   }
 
   return (
-    <div className="border-[1px] border-gray-200 w-[400px] min-h-[400px] rounded flex items-center justify-center flex-col p-4">
+    <div className="border-[1px] border-gray-200 w-[350px] min-h-[350px] rounded flex items-center justify-center flex-col p-4">
       <img
         className="w-[200px] h-[200px] rounded"
         src={image}
         onClick={() => navigate(`/detail/${id}`)}
       />
-      <p className="mt-2 min-h-[50px] text-center">{title}</p>
+      <p className="mt-2 min-h-[90px] text-center flex flex-col items-center justify-center">
+        {title}
+      </p>
       <p className="mt-2 font-bold">{price} $</p>
       <button
         onClick={handleClick}
